+++ CHANGELOG.md
@@ -414,6 +414,8 @@
 
 ## ToDo
   
+  - add to upload: greatest & least (folder sizes)
+
 ### Low Priority
 
   -> Bot